Refactor XORTrialWithLRScheduler into OneVarTrialWithLRScheduler
